First Successful Package Build
Have the basic DataM class function, as well as the dataMining() and
importantVariables() functions working properly, with their own help
files and everything. Also Description and namespace. I'm still getting
warnings when I build, but it appears to be working. Old files were
stored in the folder Temporary. The main file with the results is in
MyPackage